Remove warnings

+++ websocketpp/processors/hybi13.hpp
@@ -554,8 +554,8 @@ public:
 
         std::fill_n(
             m_extended_header.bytes,
-            frame::MAX_EXTENDED_HEADER_LENGTH,
-            0x00
+            static_cast<uint8_t>(frame::MAX_EXTENDED_HEADER_LENGTH),
+            uint8_t(0x00)
         );
     }
